updated creation of curd file
